backend/requirements.txt
fastapi
pymongo
uvicorn
python-dotenv
python-multipart
httpx
pytest
boto3